Break down buffer file
