+++ src/buffer/misc.rs
@@ -1,0 +1,123 @@
+use glam::*;
+
+use wgpu::util::DeviceExt;
+
+/// The indirect args storage buffer for [`Renderer`](crate::Renderer).
+#[derive(Debug)]
+pub struct IndirectArgsBuffer(wgpu::Buffer);
+
+impl IndirectArgsBuffer {
+    /// Create a new indirect args buffer.
+    pub fn new(device: &wgpu::Device) -> Self {
+        let buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("Indirect Args Buffer"),
+            contents: wgpu::util::DrawIndirectArgs {
+                vertex_count: 6,
+                instance_count: 0,
+                first_vertex: 0,
+                first_instance: 0,
+            }
+            .as_bytes(),
+            usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::INDIRECT,
+        });
+
+        Self(buffer)
+    }
+
+    /// Get the buffer.
+    pub fn buffer(&self) -> &wgpu::Buffer {
+        &self.0
+    }
+}
+
+/// The dispatch indirect args storage buffer for [`RadixSorter`](crate::RadixSorter).
+#[derive(Debug)]
+pub struct RadixSortIndirectArgsBuffer(wgpu::Buffer);
+
+impl RadixSortIndirectArgsBuffer {
+    /// Create a new dispatch indirect args buffer.
+    pub fn new(device: &wgpu::Device) -> Self {
+        let buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("Radix Sort Indirect Args Buffer"),
+            contents: wgpu::util::DispatchIndirectArgs { x: 1, y: 1, z: 1 }.as_bytes(),
+            usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::INDIRECT,
+        });
+
+        Self(buffer)
+    }
+
+    /// Get the buffer.
+    pub fn buffer(&self) -> &wgpu::Buffer {
+        &self.0
+    }
+}
+
+/// The indirect indices storage buffer for [`Renderer`](crate::Renderer).
+#[derive(Debug)]
+pub struct IndirectIndicesBuffer(wgpu::Buffer);
+
+impl IndirectIndicesBuffer {
+    /// Create a new indirect indices buffer.
+    pub fn new(device: &wgpu::Device, gaussian_count: u32) -> Self {
+        let buffer = device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some("Indirect Indices Buffer"),
+            size: (gaussian_count * std::mem::size_of::<u32>() as u32) as wgpu::BufferAddress,
+            usage: wgpu::BufferUsages::STORAGE,
+            mapped_at_creation: false,
+        });
+
+        Self(buffer)
+    }
+
+    /// Get the buffer.
+    pub fn buffer(&self) -> &wgpu::Buffer {
+        &self.0
+    }
+}
+
+/// The dipsatch indirect args storage buffer for [`Postprocessor`](crate::Postprocessor).
+#[derive(Debug)]
+pub struct PostprocessIndirectArgsBuffer(wgpu::Buffer);
+
+impl PostprocessIndirectArgsBuffer {
+    /// Create a new postprocessor indirect args buffer.
+    pub fn new(device: &wgpu::Device) -> Self {
+        let buffer = device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some("Postprocessor Indirect Args Buffer"),
+            size: std::mem::size_of::<wgpu::util::DispatchIndirectArgs>() as wgpu::BufferAddress,
+            usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::INDIRECT,
+            mapped_at_creation: false,
+        });
+
+        Self(buffer)
+    }
+
+    /// Get the buffer.
+    pub fn buffer(&self) -> &wgpu::Buffer {
+        &self.0
+    }
+}
+
+/// The selection storage buffer for storing selected Gaussians as a bitvec.
+#[derive(Debug)]
+pub struct SelectionBuffer(wgpu::Buffer);
+
+impl SelectionBuffer {
+    /// Create a new selection buffer.
+    pub fn new(device: &wgpu::Device, gaussian_count: u32) -> Self {
+        let buffer = device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some("Selection Buffer"),
+            size: (gaussian_count.div_ceil(32) * std::mem::size_of::<u32>() as u32)
+                as wgpu::BufferAddress,
+            usage: wgpu::BufferUsages::STORAGE,
+            mapped_at_creation: false,
+        });
+
+        Self(buffer)
+    }
+
+    /// Get the buffer.
+    pub fn buffer(&self) -> &wgpu::Buffer {
+        &self.0
+    }
+}
